+++ tests/TestCase/Shell/InstallShellTest.php
@@ -12,23 +12,17 @@
  */
 namespace MeTools\Test\TestCase\Shell;
 
+use Cake\Console\ConsoleOptionParser;
 use Cake\Http\BaseApplication;
+use Cake\Utility\Inflector;
 use MeTools\Shell\InstallShell;
 use MeTools\TestSuite\ConsoleIntegrationTestCase;
-use MeTools\TestSuite\Traits\MockTrait;
 
 /**
  * InstallShellTest class
  */
 class InstallShellTest extends ConsoleIntegrationTestCase
 {
-    use MockTrait;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $Shell;
-
     /**
      * @var array
      */
@@ -47,8 +41,6 @@ class InstallShellTest extends ConsoleIntegrationTestCase
 
         //Deletes symbolic links for plugin assets
         safe_unlink(WWW_ROOT . 'me_tools');
-
-        $this->Shell = $this->getMockForShell(InstallShell::class);
     }
 
     /**
@@ -71,8 +63,7 @@ class InstallShellTest extends ConsoleIntegrationTestCase
      */
     public function testAll()
     {
-        //Gets all methods from `InstallShell`, except for the `all()` method
-        $methods = array_diff(get_child_methods(InstallShell::class), ['all']);
+        $methods = $this->getShellMethods(['all']);
 
         $InstallShell = $this->getMockForShell(InstallShell::class, array_merge(['_stop', 'in'], $methods));
         $InstallShell->method('in')->will($this->returnValue('y'));
@@ -87,7 +78,6 @@ class InstallShellTest extends ConsoleIntegrationTestCase
         //Calls with `force` options
         $InstallShell->params['force'] = true;
         $InstallShell->all();
-
         $expectedMethodsCalledInOrder = [
             'setPermissions',
             'createRobots',
@@ -258,25 +248,12 @@ class InstallShellTest extends ConsoleIntegrationTestCase
      */
     public function testGetOptionParser()
     {
-        $this->assertEquals('Executes some tasks to make the system ready to work', $this->getParserDescription());
+        $parser = $this->Shell->getOptionParser();
+        $this->assertInstanceOf(ConsoleOptionParser::class, $parser);
+        $this->assertEquals('Executes some tasks to make the system ready to work', $parser->getDescription());
+        $this->assertArrayKeysEqual(['force', 'help', 'quiet', 'verbose'], $parser->options());
 
-        $expectedOptions = [
-            ['name' => 'force', 'short' => 'f', 'help' => 'Executes tasks without prompting'],
-            ['name' => 'help', 'short' => 'h', 'help' => 'Display this help.'],
-            ['name' => 'quiet', 'short' => 'q', 'help' => 'Enable quiet output.'],
-            ['name' => 'verbose', 'short' => 'v', 'help' => 'Enable verbose output.'],
-        ];
-        $this->assertEquals($expectedOptions, $this->getParserOptions());
-
-        $expectedSubcommands = [
-            ['name' => 'all', 'help' => 'Executes all available tasks'],
-            ['name' => 'create_directories', 'help' => 'Creates default directories'],
-            ['name' => 'create_plugins_links', 'help' => 'Creates symbolic links for plugins assets'],
-            ['name' => 'create_robots', 'help' => 'Creates the robots.txt file'],
-            ['name' => 'create_vendors_links', 'help' => 'Creates symbolic links for vendor assets'],
-            ['name' => 'fix_composer_json', 'help' => 'Fixes composer.json'],
-            ['name' => 'set_permissions', 'help' => 'Sets directories permissions'],
-        ];
-        $this->assertEquals($expectedSubcommands, $this->getParserSubcommands());
+        $expectedMethods = array_values(array_map([Inflector::class, 'underscore'], $this->getShellMethods()));
+        $this->assertArrayKeysEqual($expectedMethods, $parser->subcommands());
     }
 }